+++ anilist/types/manga.py
@@ -43,6 +43,7 @@ class Manga:
         description: str = None,
         status: str = None,
         genres: List[str] = None,
+        is_adult: bool = False,
         tags: Dict = None,
         studios: Dict = None,
         start_date: Dict = None,
@@ -79,6 +80,8 @@ class Manga:
             self.status = status
         if genres:
             self.genres = genres
+        if is_adult:
+            self.is_adult = is_adult
         if tags and len(tags) > 0:
             self.tags = [tag["name"] for tag in tags]
         if studios and len(studios["nodes"]) > 0: