+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 1.0.7 (July 25th, 2021)
+
+### Added
+
+- Message activities.
+
+### Changed
+
+- Fixed typos & bugs.
+- Added `is_adult` to both anime and manga objects.
+
 ## 1.0.6 (July 13th, 2021)
 
 ### Changed